+++ native/src/native_exports.cpp
@@ -30,7 +30,7 @@ extern "C" bool roptixRender(
     int normalFloatCount,
     const float* texCoords,
     int texCoordFloatCount,
-    const uint16_t* indices,
+    const uint32_t* indices,
     int indexCount,
     const uint32_t* triangleMaterialIndices,
     int triangleMaterialIndexCount,
