+++ src/Primitives/Cube.cs
@@ -42,7 +42,7 @@ internal static partial class Primitive {
                 0, 1,   1, 1,   1, 0,   0, 0,
             ];
 
-            ushort[] indices = [
+            uint[] indices = [
                 
                  0,  1,  2,   0,  2,  3,
                  4,  5,  6,   4,  6,  7,
